requirement.txt
absl-py==0.13.0
alembic==1.3.2
aniso8601==8.0.0
appdirs==1.4.4
astor==0.8.1
attrs==21.2.0
audioread==2.1.8
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
breadability==0.1.20
Brotli==1.0.7
bs4==0.0.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
Deprecated==1.2.13
dnspython==1.16.0
docopt==0.6.2
email-validator==1.1.1
fake-useragent==0.1.11
feedparser==6.0.2
filelock==3.0.12
filetype==1.0.8
Flask==1.1.2
Flask-Admin==1.0.8
Flask-Bcrypt==0.7.1
flask-blacklist==0.0.1
Flask-Compress==1.5.0
Flask-Cors==3.0.8
Flask-Dashboard==0.0.6a0
Flask-DebugToolbar==0.11.0
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-Moment==0.10.0
flask-mongoengine==1.0.0
Flask-MySQLdb==0.2.0
Flask-OAuth==0.12
Flask-OAuthlib==0.9.5
Flask-RESTful==0.3.7
flask-sieve==2.0.1
Flask-SQLAlchemy==2.4.4
Flask-SSLify==0.1.5
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
future==0.18.2
gast==0.2.2
geographiclib==1.52
geopy==2.2.0
GetOldTweets3==0.0.11
google-api-core==1.22.2
google-auth==1.21.1
google-auth-oauthlib==0.4.5
google-cloud-language==1.3.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
greenlet==1.1.2
grpcio==1.31.0
h5py==3.4.0
httplib2==0.19.1
idna==2.8
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.15.1
jsonschema==4.2.1
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.6.1
llvmlite==0.33.0
lxml==4.5.2
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
matplotlib==3.2.2
MechanicalSoup==0.12.0
modin==0.11.3
mongoengine==0.22.1
msgpack==1.0.2
mysqlclient==2.0.3
networkx==2.5
nltk==3.5
numba==0.47.0
numpy==1.19.0
oauth2==1.9.0.post1
oauthlib==2.1.0
opencv-python==4.4.0.40
opt-einsum==3.3.0
packaging==20.4
pandas==1.3.4
parse==1.16.0
passlib==1.7.4
peewee==3.13.3
Pillow==7.2.0
pipupgrade==1.7.4
protobuf==3.19.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
pyee==7.0.2
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
PyYAML==6.0
raven==6.10.0
ray==1.8.0
redis==4.0.2
regex==2020.7.14
requests==2.22.0
requests-html==0.10.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.5.0
selenium==3.141.0
sentencepiece==0.1.91
sgmllib3k==1.0.0
six==1.13.0
snscrape==0.3.4
SoundFile==0.10.3.post1
soupsieve==2.0.1
SQLAlchemy==1.3.23
SQLAlchemy-Searchable==1.1.0
SQLAlchemy-Utils==0.36.7
stripe==2.42.0
sumy==0.8.1
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.8.1rc1
torch==1.6.0+cpu
torchvision==0.7.0+cpu
tqdm==4.48.2
transformers==3.0.2
twitter-scraper==0.4.4
typing-extensions==3.10.0.1
urllib3==1.25.10
validators==0.15.0
w3lib==1.22.0
websockets==8.1
Werkzeug==0.16.0
wordcloud==1.7.0
wrapt==1.12.1
WTForms==2.3.1
xgboost==1.2.1
zipp==3.5.0
